Cleanup `RewardValidatorTx` logic (#1891)
